* Adjusted unit sharing and filters for Grand Alliances. (Fixes #246)
