+++ src/army/khorne/units.ts
@@ -17,39 +17,11 @@ import {
   START_OF_MOVEMENT_PHASE,
   START_OF_TURN,
 } from 'types/phases'
-import SlavesToDarkness from 'army/slaves_to_darkness'
-import Everchosen from 'army/everchosen'
-
-// Importing Khorne markable Slaves to Darkness units.
-const getSlavesUnits = () => {
-  const listOfUnits = [
-    `Daemon Prince`,
-    `Chaos Lord on Manticore`,
-    `Chaos Lord on Daemonic Mount`,
-    `Lord of Chaos`,
-    `Exalted Hero of Chaos`,
-    `Chaos Marauders`,
-    `Chaos Chariot`,
-    `Gorebeast Chariot`,
-    `Chaos Chosen`,
-    `Chaos Warriors`,
-    `Chaos Warshrine`,
-    `Chaos Knights`,
-    `Chaos Marauder Horsemen`,
-    `Chaos War Mammoth`,
-  ]
-  return SlavesToDarkness.Units.filter(unit => listOfUnits.includes(unit.name))
-}
-
-// Importing Khorne aligned Everchosen units.
-const getEverchosenUnits = () => {
-  const listOfUnits = [`Archaon`]
-  return Everchosen.Units.filter(unit => listOfUnits.includes(unit.name))
-}
+import { getSlavesUnits } from 'army/slaves_to_darkness/units'
+import { getEverchosenUnits } from 'army/everchosen/units'
 
 // Unit Names
 export const Units: TUnits = [
-  // Import Everchosen/Slaves to Darkness Units
   ...getSlavesUnits(),
   ...getEverchosenUnits(),
   {
